feat/#28: 내 정보 알림 버튼에서도 권한 체크하기
