+++ app/src/main/java/com/yeolsimee/moneysaving/view/routine/RoutineActivity.kt
@@ -2,13 +2,8 @@
 
 package com.yeolsimee.moneysaving.view.routine
 
-import android.Manifest
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.net.Uri
 import android.os.Build
 import android.os.Bundle
-import android.provider.Settings
 import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
@@ -19,9 +14,9 @@ import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.core.content.ContextCompat
 import com.yeolsimee.moneysaving.App
 import com.yeolsimee.moneysaving.domain.entity.routine.RoutineResponse
+import com.yeolsimee.moneysaving.utils.checkNotificationPermission
 import com.yeolsimee.moneysaving.utils.notification.RoutineAlarmManager
 import com.yeolsimee.moneysaving.view.category.CategoryViewModel
 import dagger.hilt.android.AndroidEntryPoint
@@ -101,7 +96,7 @@ class RoutineActivity : ComponentActivity() {
                     }
                 },
                 hasNotificationPermission = {
-                    checkNotificationPermission()
+                    checkNotificationPermission(requestPermissionLauncher)
                 },
                 onCategoryAdded = {
                     categoryViewModel.addCategory(it)
@@ -137,35 +132,5 @@ class RoutineActivity : ComponentActivity() {
         } else {
             Log.i(App.TAG, "Notification Permission Not Granted")
         }
-    }
-
-    private fun checkNotificationPermission(): Boolean {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            when {
-                ContextCompat.checkSelfPermission(
-                    this, Manifest.permission.POST_NOTIFICATIONS
-                ) == PackageManager.PERMISSION_GRANTED -> {
-                    return true
-                }
-
-                shouldShowRequestPermissionRationale(Manifest.permission.POST_NOTIFICATIONS) -> {
-                    // Responds to click on the action
-                    val intent = Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS)
-                    intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-                    val uri: Uri = Uri.fromParts("package", packageName, null)
-                    intent.data = uri
-                    startActivity(intent)
-                }
-
-                else -> {
-                    // The registered ActivityResultCallback gets the result of this request
-                    requestPermissionLauncher.launch(
-                        Manifest.permission.POST_NOTIFICATIONS
-                    )
-                }
-            }
-            return false
-        }
-        return true
     }
 }
